requirement.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
docopt==0.6.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.9
Js2Py==0.67
oauthlib==3.1.0
packaging==20.1
pipwin==0.4.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
pyparsing==2.4.6
PyPrind==2.11.2
pyttsx3==2.81
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
robobrowser==0.5.3
rsa==4.0
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
